Refactor DataContext imports to use new context path in multiple components
