+++ src/app/(checkout)/checkout/_components/new-order-form/payment-fields-group.tsx
@@ -2,7 +2,7 @@
 
 import { useFormContext } from "react-hook-form";
 
-import { useDataContext } from "@/components/DataContext";
+import { useDataContext } from "@/components/contexts/DataContext";
 import CheckboxField from "@/components/ui/form/fields/checkbox-field";
 import RadioGroupField from "@/components/ui/form/fields/radio-group-field";
 import { Text } from "@/components/ui/text";
