+++ src/app/(checkout)/payment/[uuid]/crypto/_components/payment-pending.tsx
@@ -1,4 +1,4 @@
-import { useDataContext } from "@/components/DataContext";
+import { useDataContext } from "@/components/contexts/DataContext";
 import { Button } from "@/components/ui/button";
 import CopyText from "@/components/ui/copy-text";
 import Image from "@/components/ui/image";
