+++ src/app/catalog/_components/FilterOptionButtons.tsx
@@ -3,7 +3,7 @@
 import { common_Size } from "@/api/proto-http/frontend";
 
 import { cn } from "@/lib/utils";
-import { useDataContext } from "@/components/DataContext";
+import { useDataContext } from "@/components/contexts/DataContext";
 import { Button } from "@/components/ui/button";
 
 export default function FilterOptionButtons({
