+++ src/app/template.tsx
@@ -2,7 +2,7 @@ import { serviceClient } from "@/lib/api";
 import { CartStoreProvider } from "@/lib/stores/cart/store-provider";
 import { CurrencyStoreProvider } from "@/lib/stores/currency/store-provider";
 import { LastViewedStoreProvider } from "@/lib/stores/last-viewed/store-provider.";
-import { DataContextProvider } from "@/components/DataContext";
+import { DataContextProvider } from "@/components/contexts/DataContext";
 
 export default async function Template({
   children,
